Move SomfyEntity to dedicated module
