+++ homeassistant/components/somfy/__init__.py
@@ -1,5 +1,4 @@
 """Support for Somfy hubs."""
-from abc import abstractmethod
 from datetime import timedelta
 import logging
 
@@ -8,24 +7,20 @@ import voluptuous as vol
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_CLIENT_ID, CONF_CLIENT_SECRET, CONF_OPTIMISTIC
-from homeassistant.core import HomeAssistant, callback
+from homeassistant.core import HomeAssistant
 from homeassistant.helpers import (
     config_entry_oauth2_flow,
     config_validation as cv,
     device_registry as dr,
 )
-from homeassistant.helpers.entity import Entity
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-)
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from . import api, config_flow
 from .const import API, COORDINATOR, DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
 
-SCAN_INTERVAL = timedelta(minutes=1)
+SCAN_INTERVAL = timedelta(seconds=60)
 SCAN_INTERVAL_ALL_ASSUMED_STATE = timedelta(minutes=60)
 
 SOMFY_AUTH_CALLBACK_PATH = "/auth/somfy/callback"
@@ -142,68 +137,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry):
     """Unload a config entry."""
     hass.data[DOMAIN].pop(API, None)
     return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
-
-
-class SomfyEntity(CoordinatorEntity, Entity):
-    """Representation of a generic Somfy device."""
-
-    def __init__(self, coordinator, device_id, somfy_api):
-        """Initialize the Somfy device."""
-        super().__init__(coordinator)
-        self._id = device_id
-        self.api = somfy_api
-
-    @property
-    def device(self):
-        """Return data for the device id."""
-        return self.coordinator.data[self._id]
-
-    @property
-    def unique_id(self) -> str:
-        """Return the unique id base on the id returned by Somfy."""
-        return self._id
-
-    @property
-    def name(self) -> str:
-        """Return the name of the device."""
-        return self.device.name
-
-    @property
-    def device_info(self):
-        """Return device specific attributes.
-
-        Implemented by platform classes.
-        """
-        return {
-            "identifiers": {(DOMAIN, self.unique_id)},
-            "name": self.name,
-            "model": self.device.type,
-            "via_device": (DOMAIN, self.device.parent_id),
-            # For the moment, Somfy only returns their own device.
-            "manufacturer": "Somfy",
-        }
-
-    def has_capability(self, capability: str) -> bool:
-        """Test if device has a capability."""
-        capabilities = self.device.capabilities
-        return bool([c for c in capabilities if c.name == capability])
-
-    def has_state(self, state: str) -> bool:
-        """Test if device has a state."""
-        states = self.device.states
-        return bool([c for c in states if c.name == state])
-
-    @property
-    def assumed_state(self) -> bool:
-        """Return if the device has an assumed state."""
-        return not bool(self.device.states)
-
-    @callback
-    def _handle_coordinator_update(self):
-        """Process an update from the coordinator."""
-        self._create_device()
-        super()._handle_coordinator_update()
-
-    @abstractmethod
-    def _create_device(self):
-        """Update the device with the latest data."""
